修改4烧结生产报表

+++ src/main/java/com/cisdi/steel/dto/response/sj/res/YardRunInfoOf26B.java
@@ -2,24 +2,26 @@ package com.cisdi.steel.dto.response.sj.res;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class YardRunInfoOf26B {
-    // 堆 6-天
-    private long pullMat6ForD;
-    // 堆 2-天
-    private long pushMat2ForD;
-    // 库存-天
-    private long yardInventoryBForD;
-    // 堆 6-白班
-    private long pullMat6ForShiftD;
-    // 堆 2-白班
-    private long pushMat2ForShiftD;
-    // 库存-白班
-    private long yardInventoryBForShiftD;
-    // 堆 6-夜班
-    private long pullMat6ForShiftN;
-    // 堆 2-夜班
-    private long pushMat2ForShiftN;
-    // 库存-夜班
-    private long yardInventoryBForShiftN;
+    //取料-堆6-天
+    private BigDecimal pullMat6ForD;
+    //堆料-堆2-天
+    private BigDecimal pushMat2ForD;
+    //堆区库存-b区-天
+    private BigDecimal yardInventoryBForD;
+    //取料-堆6-白班
+    private BigDecimal pullMat6ForShiftD;
+    //堆料-堆2-白班
+    private BigDecimal pushMat2ForShiftD;
+    //堆区库存-b区-白班
+    private BigDecimal yardInventoryBForShiftD;
+    //取料-堆6-夜班
+    private BigDecimal pullMat6ForShiftN;
+    //堆料-堆2-夜班
+    private BigDecimal pushMat2ForShiftN;
+    //堆区库存-b区-夜班
+    private BigDecimal yardInventoryBForShiftN;
 }
